fix: synchronize dark/light mode theme state across components
